% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pcxn.R
\name{analyze_pcxn}
\alias{analyze_pcxn}
\title{Discover correlation relationships among multiple pathways/gene sets 
identified by GSEA (gene set enrichment analysis). All the input 
pathways/gene sets should come from the same collection. MSigDB H hallmark 
gene sets, MSigDB C2 Canonical pathways, MSigDB C5 GO BP gene sets, 
and Pathprint are treated as four separate collections.}
\usage{
analyze_pcxn(collection = "pathprint",
  phenotype_0_genesets = c("ABC transporters (KEGG)",
  "ACE Inhibitor Pathway (Wikipathways)", "AR down reg. targets (Netpath)"),
  phenotype_1_genesets = c(), top = 10, min_abs_corr = 0.05,
  max_pval = 0.05)
}
\arguments{
\item{collection}{pathway's collection}

\item{phenotype_0_genesets}{genesets/pathways of the first group of pathways}

\item{phenotype_1_genesets}{genesets/pathways of the second group of pathways}

\item{top}{most correlated genesets/pathways}

\item{min_abs_corr}{minimum absolute correlation}

\item{max_pval}{maximum p-value}
}
\value{
pcxn object
}
\description{
Discover correlation relationships among multiple pathways/gene sets 
identified by GSEA (gene set enrichment analysis). All the input 
pathways/gene sets should come from the same collection. MSigDB H hallmark 
gene sets, MSigDB C2 Canonical pathways, MSigDB C5 GO BP gene sets, 
and Pathprint are treated as four separate collections.
}
\examples{
\dontrun{
 analyze_pcxn("pathprint",c("ABC transporters (KEGG)",
 "ACE Inhibitor Pathway (Wikipathways)","AR down reg. targets (Netpath)"),
  c("DNA Repair (Reactome)"), 10, 0.05, 0.05)
}
}
