% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_pcxn.R
\name{explore_pcxn}
\alias{explore_pcxn}
\title{Select a single pathway/gene set from one of the four collections ( MSigDB H 
hallmark gene sets, MSigDB C2 Canonical pathways, MSigDB C5 GO BP gene sets, 
and Pathprint ) and discover its correlated pathway/gene sets within the 
same collection.}
\usage{
explore_pcxn(collection = "pathprint",
  query_geneset = "Alzheimer's disease (KEGG)", top = 10,
  min_abs_corr = 0.05, max_pval = 0.05)
}
\arguments{
\item{collection}{pathways' collection}

\item{query_geneset}{the single pathway of interest}

\item{top}{most correlated genesets/pathways}

\item{min_abs_corr}{minimum absolute correlation}

\item{max_pval}{maximum p-value}
}
\value{
pcxn object
}
\description{
Select a single pathway/gene set from one of the four collections ( MSigDB H 
hallmark gene sets, MSigDB C2 Canonical pathways, MSigDB C5 GO BP gene sets, 
and Pathprint ) and discover its correlated pathway/gene sets within the 
same collection.
}
\examples{
\dontrun{
 explore_pcxn("pathprint","Alzheimer's disease (KEGG)", 10,
 0.05, 0.05)
}
}
