% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmembers_pcxn.R
\name{gmembers_pcxn}
\alias{gmembers_pcxn}
\title{Acquire the gene members of one of the available pathways}
\usage{
gmembers_pcxn(pathway_name = "Alzheimer's disease (KEGG)")
}
\arguments{
\item{pathway_name}{the  pathway whose members we want}
}
\value{
list of gene names and Ids
}
\description{
Acquire the gene members of one of the available pathways
}
\examples{
\dontrun{
 gmembers_pcxn("Alzheimer's disease (KEGG)")
}
}
