% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pcxn.R
\name{heatmap_pcxn}
\alias{heatmap_pcxn}
\title{Draw a heatmap of a pcxn object}
\usage{
heatmap_pcxn(df, cluster_method = "complete")
}
\arguments{
\item{df}{pcxn object created by explore or analyze functions}

\item{cluster_method}{clustering method}
}
\value{
pheatmap object
}
\description{
Draw a heatmap of a pcxn object
}
\examples{
\dontrun{
 heatmap_pcxn(df, "complete")
}
}
