% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_pcxn.R
\name{network_pcxn}
\alias{network_pcxn}
\title{Create a network of a pcxn object}
\usage{
network_pcxn(df)
}
\arguments{
\item{df}{pcxn object created by explore or analyze functions}
}
\value{
tkplot object representing the network
}
\description{
Create a network of a pcxn object
}
\examples{
\dontrun{
network <- network_pcxn(df)
}
}
