% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcxn_obj.R
\docType{class}
\name{pcxn_obj-class}
\alias{pcxn_obj-class}
\title{PCxN object which holds the type of analysis it came from, the data produced 
by the analysis and the geneset groups involved. It can be produced by the 
explore_pcxn() or analyze_pcxn() functions.}
\value{
pcxn_obj with a type, data and geneset_grous field
}
\description{
PCxN object which holds the type of analysis it came from, the data produced 
by the analysis and the geneset groups involved. It can be produced by the 
explore_pcxn() or analyze_pcxn() functions.
}
\section{Slots}{

\describe{
\item{\code{type}}{character.}

\item{\code{data}}{matrix.}

\item{\code{geneset_groups}}{list.}
}}

\examples{
\dontrun{
pcxn_obj
}

}
